tests/JG.Code.Catalog.UnitTests/Application/Genre/DeleteGenre/DeleteGenreTest.cs
using UseCase = JG.Code.Catalog.Application.UseCases.Genre.DeleteGenre;
using Moq;

namespace JG.Code.Catalog.UnitTests.Application.Genre.DeleteGenre;

[Collection(nameof(DeleteGenreTestFixture))]
public class DeleteGenreTest
{
    private readonly DeleteGenreTestFixture _fixture;

    public DeleteGenreTest(DeleteGenreTestFixture fixture)
    {
        _fixture = fixture;
    }

    [Fact(DisplayName = nameof(DeleteGenre))]
    [Trait("Application", "DeleteGenre - Use Cases")]
    public async Task DeleteGenre()
    {
        var repositoryMock = _fixture.GetGenreRepositoryMock();
        var unitOfWorkMock = _fixture.GetUnitOfWorkMock();
        var genreExample = _fixture.GetExampleGenre();
        repositoryMock.Setup(x => x.Get(genreExample.Id, It.IsAny<CancellationToken>())).ReturnsAsync(genreExample);
        var input = new UseCase.DeleteGenreInput(genreExample.Id);
        var useCase = new UseCase.DeleteGenre(repositoryMock.Object, unitOfWorkMock.Object);

        await useCase.Handle(input, CancellationToken.None);

        repositoryMock.Verify(repository => repository.Get(genreExample.Id, It.IsAny<CancellationToken>()), Times.Once);
        repositoryMock.Verify(repository => repository.Delete(genreExample, It.IsAny<CancellationToken>()), Times.Once);
        unitOfWorkMock.Verify(uow => uow.Commit(It.IsAny<CancellationToken>()), Times.Once);
    }
}
